requirements.txt
torch>=1.8.0
numpy>=1.19.0
matplotlib>=3.3.0
mysql-connector-python>=8.0.0
Flask>=2.0.0
gevent>=21.0.0
tqdm>=4.50.0